% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_dblite.R
\name{tw_dbinit}
\alias{tw_dbinit}
\title{setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.}
\usage{
tw_dbinit(dbname = "./../data/db/fm", datefrom = "1999-12-31")
}
\arguments{
\item{dbname}{a path to locate the sqlite db.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a list of objects, tables, and fields of
  balance sheet, income statement, monthly revenue, stock price, split and dividend tables.
}
\description{
setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.
}
\examples{
tw_dbinit()
}
