% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_ForeignInvestment}
\alias{tw_FM_ForeignInvestment}
\title{FM Taiwan Stock Foreign Investment.}
\usage{
tw_FM_ForeignInvestment(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ ChineseInvestmentUpperLimitRatio: num,
  $ ForeignInvestmentRemainingShares: num,
  $ ForeignInvestmentShares: num,
  $ ForeignInvestmentUpperLimitRatio: num,
  $ InternationalCode: chr,
  $ NumberOfSharesIssued: num,
  $ RecentlyDeclareDate: chr,
  $ date: chr,
  $ stock_id: chr,
  $ stock_name: chr
}
\description{
FM Taiwan Stock Foreign Investment.
}
\examples{
tw_FM_ForeignInvestment(c("2330", "2008"), "2019-01-01")
}
