% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_IncomeStatement}
\alias{tw_FM_IncomeStatement}
\title{FM Taiwan Stock Income Statement.}
\usage{
tw_FM_IncomeStatement(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ date    : chr,
  $ stock_id: chr,
  $ type    : chr,
  $ value   : num
}
\description{
FM Taiwan Stock Income Statement.
}
\examples{
tw_FM_IncomeStatement(c("2330", "2008"), "2019-01-01")
}
