% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_MonthlyRevenue}
\alias{tw_FM_MonthlyRevenue}
\title{FM Taiwan Stock Monthly Revenue.}
\usage{
tw_FM_MonthlyRevenue(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ country: chr,
  $ date: chr,
  $ revenue: num,
  $ revenue_month: int,
  $ revenue_year: int,
  $ stock_id: chr
}
\description{
FM Taiwan Stock Monthly Revenue.
}
\examples{
tw_FM_MonthlyRevenue(c("2330", "2008"), "2019-01-01")
}
