% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_StockDispersion}
\alias{tw_FM_StockDispersion}
\title{FM Taiwan Stock Dispersion.}
\usage{
tw_FM_StockDispersion(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ HoldingSharesLevel: chr,
  $ date: chr,
  $ people: int,
  $ percent: num,
  $ stock_id: chr,
  $ unit: num
}
\description{
FM Taiwan Stock Dispersion.
}
\examples{
tw_FM_StockDispersion(c("2330", "2008"), "2019-01-01")
}
