% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_StockInfo}
\alias{tw_FM_StockInfo}
\title{FM Taiwan Stock Info.}
\usage{
tw_FM_StockInfo()
}
\value{
a tibble contains following variables:
  $ industry_category: chr,
  $ stock_id: chr,
  $ stock_name: chr,
  $ type: int
}
\description{
FM Taiwan Stock Info.
}
\examples{
tw_FM_TaiwanStockInfo()
}
