% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_StockMarginShort}
\alias{tw_FM_StockMarginShort}
\title{FM Taiwan Stock Margin Trading and Short Selling.}
\usage{
tw_FM_StockMarginShort(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ MarginPurchaseBuy: int,
  $ MarginPurchaseCashRepayment: int,
  $ MarginPurchaseLimit: int,
  $ MarginPurchaseSell: int,
  $ MarginPurchaseTodayBalance: int,
  $ MarginPurchaseYesterdayBalance: int,
  $ Note: chr,
  $ OffsetLoanAndShort: int,
  $ ShortSaleBuy: int,
  $ ShortSaleCashRepayment: int,
  $ ShortSaleLimit: int,
  $ ShortSaleSell: int,
  $ ShortSaleTodayBalance: int,
  $ ShortSaleYesterdayBalance: int,
  $ date: chr,
  $ stock_id: chr,
  $ stock_name: chr
}
\description{
FM Taiwan Stock Margin Trading and Short Selling.
}
\examples{
tw_FM_StockMarginShort(c("2330", "2008"), "2019-01-01")
}
