% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_StockPrice}
\alias{tw_FM_StockPrice}
\title{FM Taiwan Stock Price.}
\usage{
tw_FM_StockPrice(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ Trading_Volume: int,
  $ Trading_money: num,
  $ Trading_turnover: int,
  $ close: num,
  $ date: chr,
  $ max: num,
  $ min: num,
  $ open: num,
  $ spread: num,
  $ stock_id: chr
}
\description{
FM Taiwan Stock Price.
}
\examples{
tw_FM_StockPrice(c("2330", "2008"), "2019-01-01")
}
