% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_FM_StockSplitDividend}
\alias{tw_FM_StockSplitDividend}
\title{FM Taiwan Stock Split and Dividend.}
\usage{
tw_FM_StockSplitDividend(symbolvec, datefrom)
}
\arguments{
\item{symbolvec}{stock symbol vector, in chr (preferred) or int type.}

\item{datefrom}{start date formatted "YYYY-MM-DD".}
}
\value{
a tibble contains following variables:
  $ Capital_Reserve: num,
  $ Cash_dividend: num,
  $ Directors_remuneration: num,
  $ Ex_dividend_transaction_day: chr,
  $ Retained_Earnings: num,
  $ date: chr,
  $ stock_id: chr,
  $ total_employee_bonus_shares: num,
  $ total_employee_bonus_stock_shares: num
}
\description{
FM Taiwan Stock Split and Dividend.
}
\examples{
tw_FM_StockSplitDividend(c("2330", "2008"), "2019-01-01")
}
