% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getFM.R
\name{tw_fmdb}
\alias{tw_fmdb}
\title{setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.}
\usage{
tw_fmdb(dbname = "./../data/db/fm")
}
\arguments{
\item{dbname}{a path to locate the sqlite db.}
}
\value{
a list of objects, tables, and fields of
  balance sheet, income statement, monthly revenue, stock price, split and dividend tables.
}
\description{
setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.
}
\examples{
tw_fmdb()
}
