% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getSec.R
\name{tw_getAllSec}
\alias{tw_getAllSec}
\title{Taiwan all securities list.}
\usage{
tw_getAllSec()
}
\value{
a tibble contains following variables: no, title, dated,
  sec_type, sec_symbol, sec_name, isin_code, start_date, end_date,
  intr_rate, market, sector, cfi_code, remark.
}
\description{
Taiwan all securities list.
}
\examples{
tw_getAllSec()
}
