% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getCM.r
\name{tw_getCMoney}
\alias{tw_getCMoney}
\title{Taiwan stock info from CMoney.}
\usage{
tw_getCMoney(symbolvec)
}
\arguments{
\item{symbolvec}{a numeric symbol vector.}
}
\value{
a list contains financial statements, ratios and monthly revenues.
}
\description{
Taiwan stock info from CMoney.
}
\examples{
tw_getCMoney(2330)
tw_getCMoney(c(2330, 2886, 3008))
}
