% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getSec.R
\name{tw_getSec}
\alias{tw_getSec}
\title{Taiwan securities list.}
\usage{
tw_getSec(stockonly = TRUE)
}
\arguments{
\item{stockonly}{a boolean value default FALSE for all securities,
if TRUE the output limited to only stocks#'}
}
\value{
a tibble contains following variables: no, title, dated,
  sec_type, sec_symbol, sec_name, isin_code, start_date, end_date,
  intr_rate, market, sector, cfi_code, remark.
}
\description{
Taiwan securities list.
}
\examples{
tw_getSec()
tw_getSec(stockonly = FALSE)
}
