% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getSec.R
\name{tw_getStocks}
\alias{tw_getStocks}
\title{Taiwan stocks list.}
\usage{
tw_getStocks()
}
\value{
a tibble contains following columns:
  market, sec_symbol, sec_name, isin_code, start_date, sector.
}
\description{
Taiwan stocks list.
}
\examples{
tw_getStocks()
}
