% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getSec.R
\name{tw_secdb}
\alias{tw_secdb}
\title{setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.}
\usage{
tw_secdb(dbname = "./../data/db/sec")
}
\arguments{
\item{dbname}{a path to locate the sqlite db.}
}
\value{
a list of objects, tables, and fields of securities
}
\description{
setup a sqlite db and populate with balance sheet, income statement, monthly revenue, stock price, split and dividend of listed stocks in twse and tpex market.
}
\examples{
tw_secdb()
}
