% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tw_getSec.R
\name{tw_tpex}
\alias{tw_tpex}
\title{Taiwan TPEX stocks list.}
\usage{
tw_tpex()
}
\value{
a tibble contains following columns:
  sec_symbol, sec_name, isin_code, start_date, sector.
}
\description{
Taiwan TPEX stocks list.
}
\examples{
tw_getStocks()
}
